Make display_order a float

+++ _INSTALL/default.sql
@@ -179,7 +179,7 @@ CREATE TABLE IF NOT EXISTS `cm_columns` (
   `select_options` text NOT NULL,
   `display_by_default` varchar(10) NOT NULL COMMENT 'Should this column be displayed to the case table user by default?',
   `required` int(11) NOT NULL DEFAULT '0' COMMENT 'ClinicCases cannot function without this field',
-  `display_order` int(11) NOT NULL,
+  `display_order` float NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `id` (`id`)
 ) ENGINE=MyISAM  DEFAULT CHARSET=latin1 COMMENT='Defines the columns to be used in ClinicCases cases table' AUTO_INCREMENT=45 ;
